refactored dataloading logic and model definitions
