+++ src/speculators/models/eagle3/__init__.py
@@ -1,7 +1,9 @@
 from speculators.models.eagle3.config import Eagle3SpeculatorConfig
 from speculators.models.eagle3.core import Eagle3DraftModel
+from speculators.models.eagle3.data import shift_batch
 
 __all__ = [
     "Eagle3DraftModel",
     "Eagle3SpeculatorConfig",
+    "shift_batch",
 ]
